+++ packages/cli/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Fixed an issue where a persistent task may not run or error with a missing dependency hash, if ran
   alongside many other persistent tasks.
+- Fixed an issue where "build" dependencies were being synced as project references.
 
 ## 1.27.0